+++ commons/dimas-commands/src/lib.rs
@@ -2,15 +2,16 @@
 
 //! Commands for `DiMAS` control & monitoring programs
 
+use chrono::Local;
 // region:		--- modules
 use dimas_com::{
-	messages::{AboutEntity, ScoutingEntity},
+	messages::{AboutEntity, PingEntity, ScoutingEntity},
 	Communicator,
 };
 use dimas_config::Config;
 use dimas_core::{
-	error::Result,
 	enums::{OperationState, Signal},
+	error::Result,
 	message_types::Message,
 };
 use itertools::Itertools;
@@ -28,7 +29,7 @@ pub fn about_list(com: &Communicator, base_selector: &String) -> Vec<AboutEntity
 	let selector = format!("{base_selector}/signal");
 	let message = Message::encode(&Signal::About);
 	// set state for entities matching the selector
-	com.get(&selector, Some(&message), |response|  -> Result<()> {
+	com.get(&selector, Some(&message), |response| -> Result<()> {
 		let response: AboutEntity = response.decode().expect("decode failed");
 		map.entry(response.zid().to_string())
 			.or_insert(response);
@@ -37,6 +38,40 @@ pub fn about_list(com: &Communicator, base_selector: &String) -> Vec<AboutEntity
 	.expect("querying 'about' failed");
 
 	let result: Vec<AboutEntity> = map.values().sorted().cloned().collect();
+
+	result
+}
+
+/// Fetch a list of about messages from all reachable `DiMAS` entities
+/// # Panics
+#[must_use]
+pub fn ping_list(com: &Communicator, base_selector: &String) -> Vec<(PingEntity, i64)> {
+	let mut map: HashMap<String, (PingEntity, i64)> = HashMap::new();
+
+	let selector = format!("{base_selector}/signal");
+	let sent = Local::now()
+		.naive_utc()
+		.and_utc()
+		.timestamp_nanos_opt()
+		.unwrap_or(0);
+	let message = Message::encode(&Signal::Ping {sent});
+	// set state for entities matching the selector
+	com.get(&selector, Some(&message), |response| -> Result<()> {
+		let received = Local::now()
+			.naive_utc()
+			.and_utc()
+			.timestamp_nanos_opt()
+			.unwrap_or(0);
+
+		let response: PingEntity = response.decode().expect("decode failed");
+		let roundtrip = received - sent;
+		map.entry(response.zid().to_string())
+			.or_insert((response, roundtrip));
+		Ok(())
+	})
+	.expect("querying 'about' failed");
+
+	let result: Vec<(PingEntity, i64)> = map.values().sorted().cloned().collect();
 
 	result
 }
@@ -76,7 +111,7 @@ pub fn set_state(
 	let selector = format!("{base_selector}/signal");
 	let message = Message::encode(&Signal::State { state });
 	// set state for entities matching the selector
-	com.get(&selector, Some(&message), |response|  -> Result<()> {
+	com.get(&selector, Some(&message), |response| -> Result<()> {
 		let response: AboutEntity = response.decode().expect("decode failed");
 		map.entry(response.zid().to_string())
 			.or_insert(response);