+++ commons/dimas-core/src/enums.rs
@@ -6,7 +6,11 @@
 // region:		--- modules
 use crate::error::{DimasError, Result};
 use bitcode::{Decode, Encode};
-use std::fmt::{Debug, Display};
+use std::{
+	fmt::{Debug, Display},
+	sync::{mpsc::Receiver, Mutex},
+	time::Duration,
+};
 // endregion:	--- modules
 
 // region:		--- OperationState
@@ -64,6 +68,11 @@ impl Display for OperationState {
 pub enum Signal {
 	/// About
 	About,
+	/// respond to Ping
+	Ping {
+		/// the utc time coordinate when the request was sent
+		sent: i64
+	},
 	/// Shutdown application
 	Shutdown,
 	/// State
@@ -73,3 +82,33 @@ pub enum Signal {
 	},
 }
 // endregion:	--- Signal
+
+// region:		--- TaskSignal
+/// Internal signals, used by panic hooks to inform that someting has happened.
+#[derive(Debug, Clone)]
+#[allow(clippy::enum_variant_names)]
+pub enum TaskSignal {
+	/// Restart a certain liveliness subscriber, identified by its key expression
+	RestartLiveliness(String),
+	/// Restart a certain queryable, identified by its key expression
+	RestartQueryable(String),
+	/// Restart a certain lsubscriber, identified by its key expression
+	RestartSubscriber(String),
+	/// Restart a certain timer, identified by its key expression
+	RestartTimer(String),
+	/// Shutdown whole process
+	Shutdown,
+}
+
+/// Wait non-blocking for [`TaskSignal`]s.<br>
+/// # Panics
+pub async fn wait_for_task_signals(rx: &Mutex<Receiver<TaskSignal>>) -> Box<TaskSignal> {
+	loop {
+		if let Ok(signal) = rx.lock().expect("snh").try_recv() {
+			return Box::new(signal);
+		};
+		// TODO: maybe there is a better solution than sleep
+		tokio::time::sleep(Duration::from_millis(1)).await;
+	}
+}
+// endregion:	--- TaskSignal