+++ commons/dimas-core/src/lib.rs
@@ -8,8 +8,6 @@ pub mod enums;
 pub mod error;
 /// `Message`, `Request`, `Response`, `Feedback`
 pub mod message_types;
-/// Task signals
-pub mod task_signal;
 /// Traits
 pub mod traits;
 /// Utilities