+++ dimas/src/com/queryable.rs
@@ -4,10 +4,9 @@
 
 // region:		--- modules
 use dimas_core::{
-	enums::OperationState,
+	enums::{OperationState, TaskSignal},
 	error::{DimasError, Result},
 	message_types::Request,
-	task_signal::TaskSignal,
 	traits::{Capability, Context},
 };
 use std::{